Inline everything in RecentActivityIcon
